Added possibility to use ipaddress_main_interface
ipaddress_main_interface is part of stdlibplus and creates a saner
logic in where the default route is taken into consideration when
figuring out which is the primary interface.
The old way would break put interface abc ahead of bdc even though
bdc has the default route set and the other interface only has an ip.